requirements.txt
flask
flask-socketio
openai
pymongo
python-dotenv